restructuring of repo
