+++ src/constants.py
@@ -1,14 +1,6 @@
 """
-Fetch env loaded secret keys, and other constants from here
+Constants from here
 """
-import os
-from dotenv import load_dotenv
-
-load_dotenv()
-DUNE_KEY = os.getenv("DUNE_KEY")
-INFURA_KEY = os.getenv("INFURA_KEY")
-ETHERSCAN_KEY = os.getenv("INFURA_KEY")
-
 ADDRESS = "0x9008D19f58AAbD9eD0D60971565AA8510560ab41"
 
 ABSOLUTE_ETH_FLAG_AMOUNT = 0.002
