--- src/quasimodo_ebbo/on_chain_surplus.py
@@ -1,308 +1,0 @@
-"""
-This component of EBBO testing parses all settlements happening on-chain and recovers
-each orders' surplus. We call Quasimodo to provide a solution for the same order,
-and then a comparison is made to determine whether the order should be flagged.
-"""
-import json
-from typing import List, Tuple, Optional, Any
-from copy import deepcopy
-import os
-from dotenv import load_dotenv
-from eth_typing import Address, HexStr
-from hexbytes import HexBytes
-import requests
-from web3 import Web3
-from src.helper_functions import (
-    get_logger,
-    get_tx_hashes_by_block,
-    get_surplus_order,
-    percent_eth_conversions_order,
-    DecodedSettlement,
-)
-from src.constants import (
-    INFURA_KEY,
-    ABSOLUTE_ETH_FLAG_AMOUNT,
-    REL_DEVIATION_FLAG_PERCENT,
-    ADDRESS,
-    SUCCESS_CODE,
-    FAIL_CODE,
-)
-from contracts.gpv2_settlement import gpv2_settlement as gpv2Abi
-
-# fetch keys
-load_dotenv()
-QUASIMODO_SOLVER_URL = os.getenv("QUASIMODO_SOLVER_URL")
-
-
-class QuasimodoTestEBBO:
-    """
-    Functions required for on-chain quasimodo calling functionality
-    """
-
-    def __init__(self) -> None:
-        """
-        Makes ethereum node connection via Infura, creates contract
-        and logger instance
-        """
-        self.web_3 = Web3(
-            Web3.HTTPProvider(f"https://mainnet.infura.io/v3/{INFURA_KEY}")
-        )
-        self.contract_instance = self.web_3.eth.contract(
-            address=Address(HexBytes(ADDRESS)), abi=gpv2Abi
-        )
-        self.logger = get_logger()
-
-    def create_tx_list(
-        self,
-        start_block: Optional[int] = None,
-        end_block: Optional[int] = None,
-        settlement_hash: Optional[str] = None,
-    ) -> None:
-        """
-        Puts transactions in list, calls required function in case of range of blocks.
-        """
-        settlement_hashes_list = []
-        if settlement_hash is not None:
-            settlement_hashes_list = [settlement_hash]
-
-        elif start_block is not None and end_block is not None:
-            settlement_hashes_list = get_tx_hashes_by_block(
-                self.web_3, start_block, end_block
-            )
-            # At this point we have all the needed hashes
-        self.decoder(settlement_hashes_list)
-
-    def decoder(self, settlement_hashes_list: List[str]) -> None:
-        """
-        Decode each hash which also calculates surplus
-        """
-        for settlement_hash in settlement_hashes_list:
-            self.process_single_hash(settlement_hash)
-
-    def process_single_hash(self, settlement_hash: str) -> bool:
-        """
-        Goes over all orders in the winning settlement, decodes orders,
-        gets response from AWS bucket, and runs the main Quasimodo test for each order.
-        Returns FALSE in case data cannot be fetched, and TRUE, otherwise.
-        """
-        try:
-            encoded_transaction = self.web_3.eth.get_transaction(
-                HexStr(settlement_hash)
-            )
-            decoded_settlement = DecodedSettlement.new(
-                self.contract_instance, encoded_transaction["input"]
-            )
-        except ValueError as except_err:
-            self.logger.error(
-                "Unhandled exception, possibly can't decode: %s", str(except_err)
-            )
-            return False
-        (winning_orders, bucket_response) = self.get_order_data_by_hash(settlement_hash)
-        if bucket_response is None:
-            return False
-        self.solve_orders_in_settlement(
-            bucket_response, winning_orders, decoded_settlement
-        )
-        return True
-
-    def get_order_data_by_hash(
-        self, settlement_hash: str
-    ) -> Tuple[List[str], Optional[Any]]:
-        """
-        Returns competition endpoint data since we need order_id and auction_id, and also AWS
-        Bucket response.
-        Note that the second tuple argument is Optinal[dict] but mypy was complaing
-        """
-        bucket_response = None
-        comp_data = None
-        try:
-            # first fetch from production environment
-            comp_data = requests.get(
-                f"https://api.cow.fi/mainnet/api/v1/solver_competition/by_tx_hash/{settlement_hash}"
-            )
-            status_code = comp_data.status_code
-            if status_code == SUCCESS_CODE:
-                compete_data = comp_data.json()
-                auction_id = compete_data["auctionId"]
-                bucket_response = dict(
-                    requests.get(
-                        (
-                            "https://solver-instances.s3.eu-central-1.amazonaws.com/"
-                            f"prod-mainnet/{auction_id}.json"
-                        )
-                    ).json()
-                )
-            elif status_code == FAIL_CODE:
-                # attempt fetch from staging environment, if prod failed.
-                comp_data = requests.get(
-                    (
-                        "https://barn.api.cow.fi/mainnet/api/v1/solver_competition/"
-                        f"by_tx_hash/{settlement_hash}"
-                    )
-                )
-                status_code = comp_data.status_code
-                if comp_data.status_code == SUCCESS_CODE:
-                    compete_data = comp_data.json()
-                    auction_id = compete_data["auctionId"]
-                    bucket_response = dict(
-                        requests.get(
-                            (
-                                "https://solver-instances.s3.eu-central-1.amazonaws.com/"
-                                f"staging-mainnet/{auction_id}.json"
-                            )
-                        ).json()
-                    )
-            return compete_data["solutions"][-1]["orders"], bucket_response
-        except ValueError as except_err:
-            self.logger.error("Unhandled exception: %s", str(except_err))
-            return [], None
-
-    @staticmethod
-    def get_solver_response(order_id: str, bucket_response: Any) -> Tuple[Any, Any]:
-        """
-        Updates AWS bucket response to a single order for posting
-        to quasimodo, in order to get the solutions JSON.
-        Note that bucket_response is a dict, but wanted to bypass type checks.
-        """
-        solver_instance = deepcopy(bucket_response)
-        order: Any = {}
-        for key, order_ in solver_instance["orders"].items():
-            if order["id"] == order_id:
-                solver_instance["orders"] = {key: order_}
-                order = order_
-                break
-        # convert back to JSON as post data
-        bucket_response_json = json.dumps(solver_instance)
-        solver_url = (
-            str(QUASIMODO_SOLVER_URL)
-            + "/solve?time_limit=20&use_internal_buffers=false&objective=surplusfeescosts"
-        )
-        # make solution request to quasimodo
-        solution = requests.post(
-            solver_url, json=bucket_response_json, timeout=30
-        ).json()
-        # return quasimodo solved solution
-        return solution, order
-
-    def solve_orders_in_settlement(
-        self,
-        bucket_response: Any,  # this is a dict, but wanted to bypass type annotation
-        winning_orders: List[str],
-        decoded_settlement: DecodedSettlement,
-    ) -> None:
-        """
-        This function goes over orders in settlement,
-        calculates surplus difference by making calls to
-        surplus calculating functions for on-chain solution, and
-        quasimodo solution.
-        """
-        # there can be multiple orders/trades in a single settlement
-        for trade in decoded_settlement.trades:
-            sell_token_clearing_price = decoded_settlement.clearing_prices[
-                trade["sellTokenIndex"]
-            ]
-            buy_token_clearing_price = decoded_settlement.clearing_prices[
-                trade["buyTokenIndex"]
-            ]
-            # convert flags value to binary to extract L.S.B (Least Sigificant Byte)
-            # 1 implies buy order, 0 implies sell order
-            order_type = str(f"{trade['flags']:08b}")[-1]
-            winning_surplus = get_surplus_order(
-                trade["executedAmount"],
-                trade["sellAmount"],
-                sell_token_clearing_price,
-                buy_token_clearing_price,
-                order_type,
-            )
-            order_id = winning_orders[decoded_settlement.trades.index(trade)]["id"]
-            solver_solution, order = self.get_solver_response(order_id, bucket_response)
-
-            sell_token = order["sell_token"]
-            buy_token = order["buy_token"]
-            # if a valid solution is returned by solver
-            if not len(solver_solution["prices"]) > 0:
-                continue
-            sell_token_clearing_price = solver_solution["prices"][sell_token]
-            buy_token_clearing_price = solver_solution["prices"][buy_token]
-            quasimodo_surplus = get_surplus_order(
-                trade["executedAmount"],
-                trade["sellAmount"],
-                sell_token_clearing_price,
-                buy_token_clearing_price,
-                order_type,
-            )
-            self.check_flag_condition(
-                winning_surplus - quasimodo_surplus,
-                trade,
-                order_type,
-                bucket_response["tokens"],
-                order,
-            )
-
-    @staticmethod
-    def check_flag_condition(
-        diff_surplus: int, trade: Any, order_type: str, tokens: Any, order: Any
-    ) -> None:
-        """
-        Based on order type, this function fetches percent_deviation,
-        and surplus difference in ETH to flag or NOT flag orders.
-        """
-        if order_type == "1":  # buy order
-            # in this case, sell amount
-            buy_or_sell_amount = int(trade["sellAmount"])
-            # external price is in the sell token for buy orders
-            external_price = int(tokens[order["sell_token"]]["external_price"])
-        elif order_type == "0":  # sell order
-            # in this case, buy amount
-            buy_or_sell_amount = int(trade["buyAmount"])
-            # external price is in the buy token for sell orders
-            external_price = int(tokens[order["buy_token"]]["external_price"])
-
-        percent_deviation, diff_in_eth = percent_eth_conversions_order(
-            diff_surplus,
-            buy_or_sell_amount,
-            external_price,
-        )
-        if (
-            percent_deviation < -REL_DEVIATION_FLAG_PERCENT
-            and diff_in_eth < -ABSOLUTE_ETH_FLAG_AMOUNT
-        ):
-            print("flag")
-
-    def print_logs(
-        self, settlement_hash: str, order_id: Any, winning_surplus: int
-    ) -> None:
-        """
-        print logs if order is flagged
-        """
-        # flag_log = "Settlement Hash: {}\nFor order: {}\nAbsolute ETH Difference: {}\n
-        # Relative Percent Difference: {}\n".format(
-        #     settlement_hash,
-        #     order_id,
-        #     str(format(diff_in_eth, ".5f")),
-        #     str(format(percent_deviation, ".4f")),
-        # )
-        flag_log = f"Settlement Hash: {settlement_hash}\nFor order: {order_id}\n \
-                        Winning surplus: {winning_surplus}\n"
-        self.logger.info(flag_log)
-
-
-# ---------------------------- TESTING --------------------------------
-def main() -> None:
-    """
-    Main loop, to be deleted.
-    """
-    instance = QuasimodoTestEBBO()
-    option_input = input("b for block input, h for hash input: ")
-    match option_input:
-        case "b":
-            start_block = input("Start Block: ")
-            end_block = input("End Block: ")
-            instance.create_tx_list(int(start_block), int(end_block), None)
-        case "h":
-            settlement_hash = input("Settlement Hash: ")
-            instance.create_tx_list(None, None, settlement_hash)
-
-
-if __name__ == "__main__":
-    main()
